rename all_repos to util
