+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode
 .env
 tokencache.json
+data
