--- client_secret.json
@@ -1,14 +1,0 @@
-{
-  "installed": {
-    "client_id": "138559140401-n5ji0c1iocef1nrfo57hibev8d3k8qca.apps.googleusercontent.com",
-    "project_id": "gh-metrics-rst",
-    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-    "token_uri": "https://oauth2.googleapis.com/token",
-    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-    "client_secret": "F1qROm9fjEUs62BY8M7iuDiD",
-    "redirect_uris": [
-      "urn:ietf:wg:oauth:2.0:oob",
-      "http://localhost"
-    ]
-  }
-}
