+++ src/metrics.rs
@@ -1,12 +1,12 @@
 use async_trait::async_trait;
 use tokio::sync::mpsc::{Receiver, Sender};
 
-mod all_repos;
 mod export_to_sheets;
 mod gql;
 mod list_repos;
 mod print;
 mod repo_participants;
+mod util;
 
 #[async_trait]
 pub trait Producer {
